requirements.txt
Django==4.1.2
django-crispy-forms==1.12.0
django-environ==0.4.5
django-extensions==3.1.3
django-allauth==0.59.0
django-otp==1.3.0
django-two-factor-auth==1.15.5
pytest-django==4.4.0
phonenumbers==8.12.37
django-ipware==3.0.7
celery==5.2.7
channels==4.0.0
django-fernet-fields==0.6
channels-redis==4.0.0
daphne==4.0.0
psycopg2-binary
python-decouple==3.6
python-dotenv==1.0.0
django-cors-headers==4.3.1
redis==4.3.4
pytest==6.2.5
wheel==0.32.3
django-tailwind==3.4.0
openai==1.6.1
python-ipware==2.0.1
tenacity==8.2.3
vertexai==0.0.1
beautifulsoup4==4.12.2
lxml==5.1.0
google-api-core
google-cloud-aiplatform
google_images_download==2.8.0
httplib2==0.22.0
Pillow==10.2.0
urllib3==2.1.0
geopy==2.4.1
pygeohash-fast==0.3.0
django_countries==7.5.1
django_import_export==3.3.7